Update build-osx.md
Updates dates, https link to OpenSSL, and instructions. We shouldn't be advising anyone to override the system OpenSSL within OS X with the --force instruction in either Homebrew or MacPorts - Doing so means that if for any reason Homebrew/MacPorts is removed later the entire system will suddenly be completely absent of OpenSSL, something that isn't desirable to say the least.